backend/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
chromadb==0.4.22
numpy==1.26.4
openai==1.12.0
PyGithub==2.1.1
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6